requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
asgiref=3.7.2=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
colorama=0.4.6=pypi_0
django=4.2.5=pypi_0
frozenlist=1.4.0=pypi_0
google-ai-generativelanguage=0.2.0=pypi_0
google-api-core=2.11.1=pypi_0
google-auth=2.22.0=pypi_0
google-generativeai=0.1.0=pypi_0
googleapis-common-protos=1.60.0=pypi_0
grpcio=1.57.0=pypi_0
grpcio-status=1.57.0=pypi_0
idna=3.4=pypi_0
libffi=3.4.4=hd77b12b_0
multidict=6.0.4=pypi_0
openai=0.28.0=pypi_0
openssl=3.0.10=h2bbff1b_2
pip=23.2.1=py311haa95532_0
proto-plus=1.22.3=pypi_0
protobuf=4.24.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
python=3.11.4=he1021f5_0
requests=2.31.0=pypi_0
rsa=4.9=pypi_0
setuptools=68.0.0=py311haa95532_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
sqlparse=0.4.4=pypi_0
tk=8.6.12=h2bbff1b_0
tqdm=4.66.1=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py311haa95532_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h8cc25b3_0
